+++ CHANGELOG.md
@@ -6,9 +6,7 @@ All notable changes to this project will be documented in this file.
 
 ### New Features
 - Call `IexecHubContract#viewDeal` instead of legacy ABI in `IexecHubAbstractService#getChainDeal`. (#24)
-- Representations of on-chain object are now immutable. (#24 #25 #26)
-### Bug Fixes
-- Add jackson annotations for `ChainReceipt` json deserialization. (#30)
+- Representations of on-chain object are now immutable. (#24 #25 #26 #30)
 ### Quality
 - Move methods to get event blocks to `iexec-core`. (#28)
 - Add `IexecHubTestService` and `Web3jTestService` classes for tests. (#29)