+++ CHANGELOG.md
@@ -10,6 +10,7 @@ All notable changes to this project will be documented in this file.
 ### Quality
 - Move methods to get event blocks to `iexec-core`. (#28)
 - Add `IexecHubTestService` and `Web3jTestService` classes for tests. (#29)
+- Load `IexecHubContract` instance only once in `IexecHubAbstractService`. (#31)
 
 ## [[2.0.1]](https://github.com/iExecBlockchainComputing/iexec-commons-poco/releases/tag/v2.0.1) 2023-05-22